decimal price field to float for elasticsearch index

+++ apps/product/models.py
@@ -25,7 +25,7 @@ class Category(BaseModel):
 class Product(BaseModel):
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='products')
     title = models.CharField(max_length=255)
-    price = models.DecimalField(max_digits=12, decimal_places=2)
+    price = models.IntegerField()
     image = models.ImageField(upload_to='products/', null=True, blank=True)
     description = models.TextField(blank=True)
     category = models.ForeignKey(Category, on_delete=models.CASCADE, related_name='products')
